refactor: vercel 배포 오류 수정 - onboarding 파일 분리
